Added full address + billing & shipping connections to checkout API
