+++ AtlasSDK/AtlasSDK/Models/Checkout.swift
@@ -10,6 +10,8 @@ public struct Checkout {
     public let cartId: String
     public let billingAddressId: String?
     public let shippingAddressId: String?
+    public let billingAddress: BillingAddress?
+    public let shippingAddress: ShippingAddress?
     public let delivery: Delivery
     public let payment: Payment?
 }
@@ -32,6 +34,8 @@ extension Checkout: JSONInitializable {
         static let cartId = "cart_id"
         static let billingAddressId = "billing_address_id"
         static let shippingAddressId = "shipping_address_id"
+        static let shippingAddress = "shipping_address"
+        static let billingAddress = "billing_address"
         static let delivery = "delivery"
         static let payment = "payment"
     }
@@ -47,6 +51,8 @@ extension Checkout: JSONInitializable {
             cartId: cartId,
             billingAddressId: json[Keys.billingAddressId].string,
             shippingAddressId: json[Keys.shippingAddressId].string,
+            billingAddress: BillingAddress(json: json[Keys.billingAddress]),
+            shippingAddress: ShippingAddress(json: json[Keys.shippingAddress]),
             delivery: delivery,
             payment: Payment(json: json[Keys.payment]))
     }
