+++ AtlasSDK/AtlasSDK/Models/Order.swift
@@ -6,29 +6,6 @@ import Foundation
 
 public struct Order {
 
-    public struct BillingAddress {
-        public let gender: Gender
-        public let firstName: String
-        public let lastName: String
-        public let street: String
-        public let additional: String?
-        public let zip: String
-        public let city: String
-        public let countryCode: String
-    }
-
-    public struct ShippingAddress {
-        public let gender: Gender
-        public let firstName: String
-        public let lastName: String
-        public let street: String?
-        public let additional: String?
-        public let zip: String
-        public let city: String
-        public let countryCode: String
-        public let pickupPoint: PickupPoint?
-    }
-
     public let orderNumber: String
     public let customerNumber: String
     public let billingAddress: BillingAddress
@@ -73,75 +50,5 @@ extension Order: JSONInitializable {
             created: NSDate(object: json[Keys.created].string, formatter: ISO8601DateFormatter),
             detailUrl: json[Keys.detailUrl].URL,
             externalPaymentUrl: json[Keys.externalPaymentUrl].URL)
-    }
-}
-
-extension Order.BillingAddress: JSONInitializable {
-
-    private struct Keys {
-        static let gender = "gender"
-        static let firstName = "first_name"
-        static let lastName = "last_name"
-        static let street = "street"
-        static let additional = "additional"
-        static let zip = "zip"
-        static let city = "city"
-        static let countryCode = "country_code"
-    }
-
-    init?(json: JSON) {
-        guard let
-            genderRaw = json[Keys.gender].string,
-            gender = Gender(rawValue: genderRaw),
-            firstName = json[Keys.firstName].string,
-            lastName = json[Keys.lastName].string,
-            street = json[Keys.street].string,
-            zip = json[Keys.zip].string,
-            city = json[Keys.city].string,
-            countryCode = json[Keys.countryCode].string else { return nil }
-        self.init(gender: gender,
-                  firstName: firstName,
-                  lastName: lastName,
-                  street: street,
-                  additional: json[Keys.additional].string,
-                  zip: zip,
-                  city: city,
-                  countryCode: countryCode)
-    }
-}
-
-extension Order.ShippingAddress: JSONInitializable {
-
-    private struct Keys {
-        static let gender = "gender"
-        static let firstName = "first_name"
-        static let lastName = "last_name"
-        static let street = "street"
-        static let additional = "additional"
-        static let zip = "zip"
-        static let city = "city"
-        static let countryCode = "country_code"
-        static let pickupPoint = "pickup_point"
-    }
-
-    init?(json: JSON) {
-        guard let
-            genderRaw = json[Keys.gender].string,
-            gender = Gender(rawValue: genderRaw),
-            firstName = json[Keys.firstName].string,
-            lastName = json[Keys.lastName].string,
-            zip = json[Keys.zip].string,
-            city = json[Keys.city].string,
-            countryCode = json[Keys.countryCode].string else { return nil }
-        let pickupPoint = PickupPoint(json: json[Keys.pickupPoint])
-        self.init(gender: gender,
-                  firstName: firstName,
-                  lastName: lastName,
-                  street: json[Keys.street].string,
-                  additional: json[Keys.additional].string,
-                  zip: zip,
-                  city: city,
-                  countryCode: countryCode,
-                  pickupPoint: pickupPoint)
     }
 }
